+++ examples/item/defaultNodes/demo/ellipse.js
@@ -17,6 +17,8 @@ const graph = new G6.Graph({
   container: 'container',
   width,
   height,
+  // translate the graph to align the canvas's center, support by v3.5.1
+  fitCenter: true,
   defaultNode: {
     type: 'ellipse',
     size: [180, 100],