+++ examples/item/defaultNodes/demo/triangle.js
@@ -14,10 +14,13 @@ const data = {
 const graph = new G6.Graph({
   container: 'container',
   width: 500,
-  height: 500, defaultNode: {
+  height: 500,
+  // translate the graph to align the canvas's center, support by v3.5.1
+  fitCenter: true,
+  defaultNode: {
     type: 'triangle',
     size: [60],
-    // 可取值：down、up
+    // options: down, up
     direction: 'up',
     style: {
       fill: '#9EC9FF',
@@ -32,23 +35,23 @@ const graph = new G6.Graph({
       position: 'bottom',
       offset: 30
     },
-    // 节点上各个方向上的链接circle配置
+    // configurations for the four linkpoints
     linkPoints: {
       top: true,
       right: true,
       bottom: true,
       left: true,
-      // circle的大小
+      // the size of the linkpoints' circle
       size: 5,
       lineWidth: 1,
       fill: '#fff',
       stroke: '#1890FF',
     },
-    // 节点中icon配置
+    // configurations for the icon
     icon: {
-      // 是否显示icon，值为 false 则不渲染icon
+      // whether to show the icon
       show: true,
-      // icon的地址，字符串类型
+      // the image url for the icon, string type
       img:
         'https://gw.alipayobjects.com/zos/basement_prod/012bcf4f-423b-4922-8c24-32a89f8c41ce.svg',
       width: 40,
@@ -59,11 +62,11 @@ const graph = new G6.Graph({
     default: ['drag-canvas', 'drag-node'],
   },
   nodeStateStyles: {
-    // 鼠标hover状态下的配置
+    // style configurations for the hover state
     hover: {
       fillOpacity: 0.8,
     },
-    // 选中节点状态下的配置
+    // style configurations for the selected state
     selected: {
       lineWidth: 5,
     },